feat: agregados campos de refinanciamiento
